% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot_functions.R
\name{base_spp_plot}
\alias{base_spp_plot}
\title{Base Plots: Species}
\usage{
base_spp_plot(data, plotMain = NULL, colorVar = NULL, colorLegend = "none")
}
\arguments{
\item{data}{formatted data from get_spp_data()}

\item{plotMain}{main title for the figure}

\item{colorVar}{variable for the color variation. Color by species or by niche optima}

\item{colorLegend}{should the color legend be included, what is the title?}
}
\description{
Base species plot
}
