% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{C_f}
\alias{C_f}
\title{Metacommunity process: colonization}
\usage{
C_f(v, d_c, c_0, c_max)
}
\arguments{
\item{v}{sum of interactions}

\item{d_c}{sensitivity to interactions}

\item{c_0}{colonization at zero interactions}

\item{c_max}{colonization at max interactions}
}
\description{
The total effect of ecological interactions on the colonization probabilty
}
