% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{E_f}
\alias{E_f}
\title{Metacommunity process: interactions and extinction}
\usage{
E_f(v, d_e, e_0, e_min)
}
\arguments{
\item{v}{sum of ecological interactions}

\item{d_e}{sensitivity to interactions}

\item{e_0}{extinction at zero interactions}

\item{e_min}{extinction at max interactions}
}
\description{
Effect of ecological interactions on species extinction at the patch level. In this case, extinction probability must be larger when interactions are negative and smaller when the interactions are positive.
}
