% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{get_E}
\alias{get_E}
\title{Environmental variables}
\usage{
get_E(D, N)
}
\arguments{
\item{D}{Number of environmental packages}

\item{N}{Number of patches}
}
\description{
Generate random environmental variables
}
