% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot_functions.R
\name{get_fig2_params}
\alias{get_fig2_params}
\title{Get parameters for Figure 2}
\usage{
get_fig2_params(folderpath, scenario)
}
\arguments{
\item{folderpath}{path to stored RDS files from the get_VP_results()}

\item{scenario}{the specific name of the file to use. Should be able to access the both the metacomSim, the model and the VP results with this name.}
}
\description{
Manuscript specific. This function is for Figure 2 only.It takes the RDS params file and organizes it from a list to a dataframe, so that it can be joined to the Variation Partition results at the species level. That way we can know what niche values each species has. The output of this function can be joined to the VP results at the species level and then plot.
}
