% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot_functions.R
\name{get_fig3_params}
\alias{get_fig3_params}
\title{Get parameters for Figure 3}
\usage{
get_fig3_params(folderpath, scenario)
}
\arguments{
\item{folderpath}{path to stored RDS files from the get_VP_results()}

\item{scenario}{the specific name of the file to use. Should be able to access the both the metacomSim, the model and the VP results with this name.}
}
\description{
Manuscript specific, figure 3 only.Figure 3 has three sets of parameters because we divided simulations into groups.Half of the species in each simulation has interactions, and the other half doesn't.Or, a third of the species is assigned a different dispersal level.
}
