% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{get_K}
\alias{get_K}
\title{Metacommunity process: patch connectivity}
\usage{
get_K(XY, alpha)
}
\arguments{
\item{XY}{corresponds to the patch coordinates}

\item{alpha}{dispersal parameter associated to an exponential distribution used for dispersal}
}
\description{
This function will calculate the connectivity matrix for a given set of patches and a dispersal parameter
}
\keyword{patch}
