% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{get_mcmc_lists}
\alias{get_mcmc_lists}
\title{Gelman Diagnostics for all chains}
\usage{
get_mcmc_lists(folderpath = NULL, scenario = NULL)
}
\arguments{
\item{folderpath}{location of files}

\item{scenario}{specific file name}
}
\description{
Get all the mcmc objects in a list for gelman.diag
}
