% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{get_mcmc_obj}
\alias{get_mcmc_obj}
\title{Get MCMC object}
\usage{
get_mcmc_obj(folderpath = NULL, scenario = NULL, modelNum = NULL)
}
\arguments{
\item{folderpath}{location of the file}

\item{scenario}{file name}

\item{modelNum}{select from the HMSC model object}
}
\description{
Using as.mcmc function from HMSC isn't working to include the burn section of the results
}
