% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VariPart_data.R
\name{get_sites_data}
\alias{get_sites_data}
\title{Organize and format site level data}
\usage{
get_sites_data(folderpath, scenario)
}
\arguments{
\item{folderpath}{path to stored RDS files from the get_VP_results()}

\item{scenario}{the specific name of the file to use. Should be able to access the both the metacomSim, the model and the VP results with this name.}
}
\description{
Manuscript specific function. This function takes the simulated metacommunity to get the species richness for each site. It then binds that information to the Variation Partinioning output from the HMSC::variPart(). This function also takes the output of HMSC::variPart() and organizes each fraction: into the env, spa and codist components
}
