% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VariPart_data.R
\name{get_species_data}
\alias{get_species_data}
\title{Organize and Format VP results at the species level}
\usage{
get_species_data(folderpath, scenario)
}
\arguments{
\item{folderpath}{path to stored RDS files from the get_VP_results()}

\item{scenario}{the specific name of the file to use. Should be able to access the both the metacomSim, the model and the VP results with this name.}
}
\description{
this function organizes species level data and gets it ready to plot. It takes the RDS file that is the output of the HMSC::variPart() function and reorganizes it into the env, codist and spa components.It also calculates the prevalence information for each species.
organizes each fraction: into the env, spa and codist components
}
