% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variPart_functions.R
\name{get_VPresults}
\alias{get_VPresults}
\title{Variation partinioning: species}
\usage{
get_VPresults(
  HMSCmodel,
  MEMsel,
  numClusters,
  makeRDS = FALSE,
  whereToSave = NULL,
  objName = NULL
)
}
\arguments{
\item{HMSCmodel}{output from the metacom_as_HMSCdata() or an object from as.HMSCdata()}

\item{MEMsel}{distance-based Moran' Eigenvector Map's eigenvector maps from a geographic distance matrix (xy matrix). This comes from adespatial::dbmem()}

\item{numClusters}{number of clusters based on number of cores available to work with DoParallel}

\item{makeRDS}{should the function make an RDS file from the output and save it? Default is FALSE}

\item{whereToSave}{file path for the RDS file to be saved in}

\item{objName}{should the RDS file be saved, what should it be called?}
}
\description{
use of the VariPart function from HMSC, gets results at the level of species
}
