% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{get_xy_agg}
\alias{get_xy_agg}
\title{Aggregated XY patches}
\usage{
get_xy_agg(N, Nclusters, sd_xy)
}
\arguments{
\item{N}{number of patches}

\item{Nclusters}{number of clusters of patches}

\item{sd_xy}{separation between patches inside clusters}
}
\description{
Generates random set of coordinates for aggregated patches
}
