% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{I_f}
\alias{I_f}
\title{Metacommunity process: immigration}
\usage{
I_f(
  Y = "Species occurrence matrix",
  K = "Patch connectivity matrix",
  m = "outside immigration"
)
}
\arguments{
\item{Y}{matrix of species occurrence}

\item{K}{connectivity matrix}

\item{m}{independent constant to account for immigration from outside the simulated metacommunity}
}
\description{
This function will calculate the propagule pressure. The effects of immigration are given as a weighted average of the occurrence probability of species i in neighborhood of z.
}
\keyword{immigration}
\keyword{metacommunity}
