% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_correlation_matrices.R
\name{interaction_plot}
\alias{interaction_plot}
\title{Create the species correlation matrices}
\usage{
interaction_plot(folderpath, scenario, iteration = NULL,
  corTitle = NULL)
}
\arguments{
\item{folderpath}{path to stored RDS files from the get_VP_results()}

\item{scenario}{the specific name of the file to use. Should be able to access the HMSC model.}

\item{iteration}{Which iteration of the model should it use? Default is 1}

\item{corTitle}{Title for the correlation plot}
}
\description{
This mostly follows the code provided from the HMSC vignette, therefore it is not in a tidyverse framework and requires the corrplot package
}
