% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{M_f}
\alias{M_f}
\title{Metacomminbut process: extinction}
\usage{
M_f(E, u_e, s_e)
}
\arguments{
\item{E}{Matrix of environmental covariates at each patch}

\item{u_e}{extinction rate at environmental optimum}

\item{s_e}{extintion rate at minimum}
}
\description{
Responses of the extinction probability to the local environment. Extinction rate should be minimal at environmental optimum
}
