% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_simulation.R
\name{mainfx}
\alias{mainfx}
\title{Metacommunity Simulation}
\usage{
mainfx(XY, E, pars, Y0, nsteps, envResp = "quadratic")
}
\arguments{
\item{XY}{coordinates for each of the patches}

\item{E}{environmental covariates measured for each patch. Our simulation only included one so far}

\item{pars}{output from the prep_pars function. It is a list of necessary parameters to run all the process functions}

\item{Y0}{matrix of initial conditions, occurrence matrix}

\item{nsteps}{time steps to simulate metacomminuty dynamics}

\item{envResp}{choose between a "quadratic" or "gaussian" response to the environment. Default is "quadratic".}
}
\description{
This function will use all the process functions to generate a metacommunity with the given parameters. It will only store the occurrence matrix at the end of the simulation
}
