% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mce2-package.R
\docType{package}
\name{mce2-package}
\alias{mce2}
\alias{mce2-package}
\title{mce2: Metacommunity simulation and analysis version 2}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/javirudolph/mce2}
  \item Report bugs at \url{https://github.com/javirudolph/mce2/issues}
}

}
\author{
\strong{Maintainer}: Javiera Rudolph \email{first.last@example.com} (\href{https://orcid.org/https://orcid.org/0000-0002-7124-4649}{ORCID})

}
\keyword{internal}
