% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacom_as_HMSCdata.R
\name{metacom_as_HMSCdata}
\alias{metacom_as_HMSCdata}
\title{Format metacommunity data to "HMSC" classes}
\usage{
metacom_as_HMSCdata(
  metacomData,
  numClusters,
  E,
  MEMsel,
  hmscPars = NULL,
  makeRDS = FALSE,
  whereToSave = NULL,
  objName = NULL
)
}
\arguments{
\item{metacomData}{list of simulated metacommunity matrices. Output from metacom_sim4HMSC() or metacom_sim4HMSC_multiparam()}

\item{numClusters}{number of clusters based on number of cores available to work with DoParallel}

\item{E}{matrix of environmental variables associated to each site}

\item{MEMsel}{distance-based Moran' Eigenvector Map's eigenvector maps from a geographic distance matrix (xy matrix). This comes from adespatial::dbmem()}

\item{hmscPars}{list of parameters to be included for iterations, burn and thin components of the as.HMSCdata()}

\item{makeRDS}{should the function make an RDS file from the output and save it? Default is FALSE}

\item{whereToSave}{file path for the RDS file to be saved in}

\item{objName}{should the RDS file be saved, what should it be called?}
}
\description{
This function will use the simulated metacommunity matrices and format the data into "HMSC" classes
}
