% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacom_sim4HMSC.R
\name{metacom_sim4HMSC}
\alias{metacom_sim4HMSC}
\title{Metacommunity Simulation one set of parameters}
\usage{
metacom_sim4HMSC(XY, E, pars, nsteps, occupancy, niter,
  envResp = "quadratic", makeRDS = FALSE, whereToSave = NULL,
  objName = NULL)
}
\arguments{
\item{XY}{coordinates for each of the sites or patches}

\item{E}{matrix of environmental variables measured at each site}

\item{pars}{list of parameters given by the output of the prep_pars() function}

\item{nsteps}{numeric, number of time steps before getting the "snapshot" of a metacommunity}

\item{occupancy}{numeric value between 0-1, to set as the initial conditions occupancy of sites by the different species.}

\item{niter}{number of iterations for simulating the metacommunity with the same parameters.}

\item{envResp}{type of response to the environment: "gaussian" or "quadratic"}

\item{makeRDS}{should the function make an RDS file from the output and save it? Default is FALSE}

\item{whereToSave}{file path for the RDS file to be saved in}

\item{objName}{should the RDS file be saved, what should it be called?}
}
\description{
This function will use the set of given parameters and run the main metacommunity simulation for a set amount of time steps. The output is a list of matrices, where each one of the matrices is an iteration of the metacommunity with the same parameters. It gives the option to save RDS files to the specified directory so these RDS files can be accessed later by other HMSC functions.
}
