% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pars.R
\name{prep_pars}
\alias{prep_pars}
\title{Prepare metacommunity parameters}
\usage{
prep_pars(
  N = NULL,
  D = NULL,
  R = NULL,
  nicheOpt = NULL,
  breadth = NULL,
  alpha = NULL,
  interx_col = NULL,
  interx_ext = NULL,
  makeRDS = FALSE,
  whereToSave = NULL,
  objName = NULL
)
}
\arguments{
\item{N}{number of patches for the simulation}

\item{D}{number of environmental variables simulated}

\item{R}{number of species for the simulation}

\item{breadth}{niche breadth. As reference 0.8 as narrow and 2 as broad with a quadratic response.}

\item{alpha}{dispersal parameter associated to an exponential distribution}

\item{interex}{Species sensitivity to interactions. For our simulations, sensitivity is set the same for all species, and both for colonization and extinction}

\item{initocup}{initial occupancy. For our simulations set to 0.8}
}
\description{
This function is used to prepare the parameters in a useful format and avoid lengthy code for every simulation. It will output all parameters in a list ready for the main_sim_fx()
}
