% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaco_processes.R
\name{S_f_gaussian}
\alias{S_f_gaussian}
\title{Metacommunity process: Environmental filtering with gaussian response}
\usage{
S_f_gaussian(E, u_c, s_c)
}
\arguments{
\item{E}{matrix of environmental covariates for each patch}

\item{u_c}{considered as the niche optima for each species, for each environmental variable.}

\item{s_c}{understood as niche breadth for each species for each environmental variable.}
}
\description{
The effect of the environment on the local performance and colonization. Probability of establishing a viable local population given environmental conditions in the patch.
}
\keyword{environmental}
\keyword{filtering}
\keyword{gaussian}
\keyword{response}
