% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mce2-package.R
\docType{package}
\name{mce2-package}
\alias{mce2}
\alias{mce2-package}
\title{mce2: Metacommunity simulation and analysis version 2}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/javirudolph/mce2}
  \item Report bugs at \url{https://github.com/javirudolph/mce2/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
